+++ src/ElasticSearch/SearchFactory.php
@@ -6,6 +6,7 @@ use Laravel\Scout\Builder;
 use ONGR\ElasticsearchDSL\Query\Compound\BoolQuery;
 use ONGR\ElasticsearchDSL\Query\FullText\QueryStringQuery;
 use ONGR\ElasticsearchDSL\Query\TermLevel\TermQuery;
+use ONGR\ElasticsearchDSL\Query\TermLevel\TermsQuery;
 use ONGR\ElasticsearchDSL\Search;
 use ONGR\ElasticsearchDSL\Sort\FieldSort;
 
@@ -20,11 +21,10 @@ final class SearchFactory
     {
         $search = new Search();
         $query = new QueryStringQuery($builder->query);
-        if (! empty($builder->wheres)) {
+        if (static::hasWhereFilters($builder)) {
             $boolQuery = new BoolQuery();
-            foreach ($builder->wheres as $field => $value) {
-                $boolQuery->add(new TermQuery((string) $field, $value), BoolQuery::FILTER);
-            }
+            $boolQuery = static::addWheres($builder, $boolQuery);
+            $boolQuery = static::addWhereIns($builder, $boolQuery);
             $boolQuery->add($query, BoolQuery::MUST);
             $search->addQuery($boolQuery);
         } else {
@@ -43,5 +43,64 @@ final class SearchFactory
         }
 
         return $search;
+    }
+
+    /**
+     * @param Builder $builder
+     * @return bool
+     */
+    private static function hasWhereFilters($builder): bool
+    {
+        return static::hasWheres($builder) || static::hasWhereIns($builder);
+    }
+
+    /**
+     * @param Builder $builder
+     * @param BoolQuery $boolQuery
+     * @return BoolQuery
+     */
+    private static function addWheres($builder, $boolQuery): BoolQuery
+    {
+        if (static::hasWheres($builder)) {
+            foreach ($builder->wheres as $field => $value) {
+                $boolQuery->add(new TermQuery((string) $field, $value), BoolQuery::FILTER);
+            }
+        }
+
+        return $boolQuery;
+    }
+
+    /**
+     * @param Builder $builder
+     * @param BoolQuery $boolQuery
+     * @return BoolQuery
+     */
+    private static function addWhereIns($builder, $boolQuery): BoolQuery
+    {
+        if (static::hasWhereIns($builder)) {
+            foreach ($builder->whereIns as $field => $arrayOfValues) {
+                $boolQuery->add(new TermsQuery((string) $field, $arrayOfValues), BoolQuery::FILTER);
+            }
+        }
+
+        return $boolQuery;
+    }
+
+    /**
+     * @param Builder $builder
+     * @return bool
+     */
+    private static function hasWheres($builder): bool
+    {
+        return ! empty($builder->wheres);
+    }
+
+    /**
+     * @param Builder $builder
+     * @return bool
+     */
+    private static function hasWhereIns($builder): bool
+    {
+        return isset($builder->whereIns) && ! empty($builder->whereIns);
     }
 }